+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 * Fixed closing of `database/sql` connection (aka `YDB` session)
 * Made `session.Close()` as `nop` for idled session
 * Implemented goroutine for closing idle connection in `database/sql` driver
+* Separated errors of commit from other reader and to expired session
+* Fixed wrapping error in `internal/balancer/Balancer.wrapCall()`
 
 ## v3.42.4
 * Added `ydb.WithDisableServerBalancer()` database/sql connector option